# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IDEABOX SWAGGER",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/generate-text": {
            "post": {
                "tags": [
                    "GPT"
                ],
                "summary": "Generate text using GPT-3.5",
                "operationId": "ee300fa76206eeee129153f06d2a579c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text",
                                    "category_id"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Your text here"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Generated text successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "generated_text": {
                                                    "type": "string",
                                                    "example": "Generated text here"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_page": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Category ID is missing"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_page": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "API key is missing"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_page": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Category not found"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_page": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Failed to generate text"
                                                },
                                                "errors": {
                                                    "type": "string",
                                                    "example": "Error message here"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_page": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favourites/idea": {
            "post": {
                "tags": [
                    "Favourites"
                ],
                "summary": "Create a new favourite",
                "description": "Insert a new favourite into the database",
                "operationId": "addFav",
                "requestBody": {
                    "description": "Favourite data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value",
                                    "name",
                                    "category_id"
                                ],
                                "properties": {
                                    "value": {
                                        "type": "string",
                                        "example": "Some value"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Favourite Name"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Favourite created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Favourites"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Favourites": {
                "properties": {
                    "id": {
                        "description": "The unique identifier for the favourite",
                        "type": "integer",
                        "format": "int64"
                    },
                    "generate_date": {
                        "description": "The date when the favourite was generated",
                        "type": "string",
                        "format": "date"
                    },
                    "value": {
                        "description": "The value of the favourite",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the favourite",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "The ID of the user who created the favourite",
                        "type": "integer",
                        "format": "int64"
                    },
                    "category_id": {
                        "description": "The ID of the category to which the favourite belongs",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "The date and time when the favourite was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the favourite was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}